Moved FWCore/ServiceRegistry tests to catch2
